requirements.txt
Flask==2.3.3
PyPDF2==3.0.1
python-docx==0.8.11
spacy==3.7.2
scikit-learn==1.3.0
numpy==1.24.3
scipy==1.10.1
joblib==1.3.2
threadpoolctl==3.2.0
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3